notification-worker/requirements.txt
django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.9
pika==1.3.2
python-decouple==3.8
requests==2.31